pyproject.toml
[project]
name = "browser-use-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.1.32",
    "gradio>=5.14.0",
    "playwright>=1.49.1",
    "python-dotenv>=1.0.1",
    "google-generativeai>=0.3.2",
    "langchain-google-genai>=0.0.5",
    "google-cloud-aiplatform>=1.40.0",
    "streamlit>=1.30.0",
    "Pillow>=10.1.0",
    "ruff>=0.0.0"
]

[tool.ruff]
line-length = 88
exclude = ["__pycache__", "*.pyc"]
